week_2_hydra_config/requirements.txt
pytorch-lightning==1.2.10
datasets==1.6.2
transformers==4.5.1
scikit-learn==0.24.2
wandb
torchmetrics
matplotlib
seaborn
hydra-core==1.1.0
omegaconf
hydra_colorlog
